% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect_dataset}
\alias{connect_dataset}
\alias{connect_speciesgrids_local}
\alias{connect_speciesgrids}
\alias{connect_opendata_local}
\alias{connect_opendata}
\alias{connect_obistherm_local}
\alias{connect_obistherm}
\title{Connect to a specific OBIS dataset}
\usage{
connect_speciesgrids_local(path, ...)

connect_speciesgrids(...)

connect_opendata_local(path, ...)

connect_opendata(...)

connect_obistherm_local(path, ...)

connect_obistherm(...)
}
\arguments{
\item{path}{path to a local copy of the dataset (see \code{\link[=sync_speciesgrids]{sync_speciesgrids()}},
\code{\link[=sync_opendata]{sync_opendata()}}, \code{\link[=sync_obistherm]{sync_obistherm()}}).}

\item{...}{further arguments passed to \code{\link[=connect]{connect()}} (\code{select}/\code{where}).}
}
\value{
a \code{robisdb_conn} object.
}
\description{
Thin wrappers around \code{\link[=connect]{connect()}} for each of the three OBIS datasets it supports, in
their local (\verb{_local}, given a \code{path} to a synced copy) and remote (S3-streamed) forms.
}
